[IMP] Refactor code
1. Remove moron invoice, use invoice of odoo
2. Refactor code
3. imp view
