+++ local-addons/morons/__init__.py
@@ -1,6 +1,1 @@
-# -*- coding: utf-8 -*-
-
-from . import controllers
 from . import models
-
-# admin can see all contributors and all POs that they are creating: old/ current and upcoming
