+++ local-addons/morons/data/sequence.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="increment_project_id_sequence" model="ir.sequence">
+        <field name="name">Project ID Increment</field>
+        <field name="code">increment_project_id</field>
+        <field name="prefix">P</field>
+        <field name="padding">5</field>
+        <field name="company_id" eval="False" />
+    </record>
+</odoo>
