--- local-addons/morons/views/invoice.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <data>
-        <!-- <record id="invoice_access_all_users" model="ir.model.access">
-            <field name="name">invoice access for all users</field>
-            <field name="model_id" ref="morons.model_invoice" />
-            <field name="group_id" ref="base.group_user" />
-            <field name="perm_read" eval="True" />
-            <field name="perm_write" eval="True" />
-            <field name="perm_create" eval="True" />
-            <field name="perm_unlink" eval="True" /> -->
-
-    </data>
-</odoo>
