+++ local-addons/morons/views/task.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <!--project.view_task_form2-->
+    <!-- TASK VIEW-->
+    <record id="merctrans_task_view_form" model="ir.ui.view">
+        <field name="name">merctrans.task.form.view.inherit</field>
+        <field name="model">project.task</field>
+        <field name="inherit_id" ref="project.view_task_form2" />
+        <field name="arch" type="xml">            
+            <xpath expr="//field[@name='project_id']" position="attributes">
+                <attribute name="invisible">True</attribute>
+            </xpath>
+            <xpath expr="//field[@name='user_ids']" position="after">
+                <field name="rate" />
+                <field name="service" widget="many2many_tags" force_save="1"/>
+                <field name="source_language" force_save="1"/>
+                <field name="target_language" widget="many2many_tags" force_save="1"/>
+                <field name="work_unit" force_save="1"/>
+                <field name="stages_id"  />                   
+            </xpath>
+            <xpath expr="//field[@name='partner_id']" position="after">
+                <field name="show_po_value_currency_string" invisible="1"/>
+                <field name="volume" />
+                <label for="po_value" />
+                <div name="po_value_div" class="o_row">
+                    <field name="po_value" class="oe_inline" />
+                    <field name="po_value_by_project_currency_string" class="oe_inline" attrs="{'invisible': [('show_po_value_currency_string', '=', False)]}" />
+                </div>
+                <field name="currency_id" />
+                <field name="payment_status" />
+                <field name="project_currency_id" invisible="1"/>
+            </xpath>
+            <xpath expr="//field[@name='partner_id']" position="attributes">
+                <attribute name="invisible">True</attribute>
+            </xpath>
+            <xpath expr="//field[@name='tag_ids']" position="attributes">
+                <attribute name="invisible">True</attribute>
+            </xpath>
+        </field>
+    </record>
+</odoo>
